+++ Client/src/models/languageLocalisation/loginMenu.java
@@ -1,4 +1,4 @@
-package models.languages;
+package models.languageLocalisation;
 
 public class loginMenu{
     public String WELCOME_MESSAGE;
